# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stabbur API",
    "description": "Software artifact control plane and outbound worker coordination API",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/bootstrap": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "bootstrap",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "First administrator created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "410": {
            "description": "Bootstrap unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/operations/status": {
      "get": {
        "tags": [
          "operations"
        ],
        "operationId": "operational_status",
        "responses": {
          "200": {
            "description": "Operational measurements",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/releases/{release}/withdraw": {
      "post": {
        "tags": [
          "releases"
        ],
        "operationId": "withdraw_release",
        "parameters": [
          {
            "name": "release",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Release withdrawn; lifecycle and bytes preserved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseResponse"
                }
              }
            }
          },
          "409": {
            "description": "Already withdrawn",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "412": {
            "description": "Stale revision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/software/{software}/status": {
      "get": {
        "tags": [
          "software"
        ],
        "operationId": "software_status",
        "parameters": [
          {
            "name": "software",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Software execution and publication status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoftwareStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/workers/{worker}/drain": {
      "post": {
        "tags": [
          "workers"
        ],
        "operationId": "drain_worker",
        "parameters": [
          {
            "name": "worker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrainWorkerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Worker claim policy updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerResponse"
                }
              }
            }
          },
          "412": {
            "description": "Stale revision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BootstrapRequest": {
        "type": "object",
        "description": "First-administrator request using the one-time secret file contents.",
        "required": [
          "secret",
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "format": "password",
            "description": "First administrator password."
          },
          "secret": {
            "type": "string",
            "format": "password",
            "description": "One-time bootstrap secret."
          },
          "username": {
            "type": "string",
            "description": "First administrator login name."
          }
        }
      },
...
      "DrainWorkerRequest": {
        "type": "object",
        "description": "Change whether a worker may claim new work while preserving active leases.",
        "required": [
          "draining"
        ],
        "properties": {
          "draining": {
            "type": "boolean",
            "description": "True pauses new claims; false resumes claims."
          }
        },
        "additionalProperties": false
      },
...
      "OperationalStatusResponse": {
        "type": "object",
        "description": "Durable queue and worker measurements; contains no backend paths or credentials.",
        "required": [
          "queued_jobs",
          "running_jobs",
          "failed_jobs",
          "expired_attempts",
          "draining_workers"
        ],
        "properties": {
          "draining_workers": {
            "type": "integer",
            "format": "int64",
            "description": "Enabled workers paused for maintenance.",
            "minimum": 0
          },
          "expired_attempts": {
            "type": "integer",
            "format": "int64",
            "description": "Historical expired attempts.",
            "minimum": 0
          },
          "failed_jobs": {
            "type": "integer",
            "format": "int64",
            "description": "Historical terminal job failures.",
            "minimum": 0
          },
          "oldest_queued_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Creation time of the oldest queued job."
          },
          "queued_jobs": {
            "type": "integer",
            "format": "int64",
            "description": "Jobs awaiting a worker.",
            "minimum": 0
          },
          "running_jobs": {
            "type": "integer",
            "format": "int64",
            "description": "Jobs with a leased attempt.",
            "minimum": 0
          }
        }
      },
...
      "PrincipalResponse": {
        "type": "object",
        "description": "Safe authenticated principal representation.",
        "required": [
          "id",
          "name",
          "kind",
          "roles"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "UUIDv7 identity."
          },
          "kind": {
            "type": "string",
            "description": "`human`, `service`, or `worker`."
          },
          "name": {
            "type": "string",
            "description": "Login or service name."
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Assigned role names."
          }
        }
      },
      "Problem": {
        "type": "object",
        "description": "Stable `application/problem+json` response body.",
        "required": [
          "code",
          "status",
          "detail",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable Stabbur code."
          },
          "detail": {
            "type": "string",
            "description": "Human-readable detail safe for untrusted clients."
          },
          "request_id": {
            "type": "string",
            "description": "Request correlation identity."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status.",
            "minimum": 0
          },
          "validation_errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "description": "Field-level validation failures."
          }
        }
      },
...
      "RejectReleaseRequest": {
        "type": "object",
        "description": "Explicit release rejection request.",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Required safe operator reason."
          }
        }
      },
      "ReleaseAvailabilityResponse": {
        "oneOf": [
          {
            "type": "object",
            "description": "The release may be published if its verification/lifecycle permits it.",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "available"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "The release was withdrawn and cannot be promoted or resolved.",
            "required": [
              "reason",
              "at",
              "kind"
            ],
            "properties": {
              "at": {
                "type": "string",
                "format": "date-time",
                "description": "Withdrawal timestamp."
              },
              "kind": {
                "type": "string",
                "enum": [
                  "withdrawn"
                ]
              },
              "reason": {
                "type": "string",
                "description": "Operator explanation."
              }
            }
          }
        ],
        "description": "Publication eligibility of an immutable release."
      },
...
      "ReleaseResponse": {
        "type": "object",
        "description": "One immutable release with its highest achieved lifecycle state.",
        "required": [
          "availability",
          "id",
          "software_id",
          "version",
          "state",
          "created_at",
          "revision"
        ],
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/ReleaseAvailabilityResponse",
            "description": "Publication eligibility independent of attained lifecycle."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation time."
          },
          "id": {
            "type": "string",
            "description": "UUIDv7 release identity."
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "description": "Optimistic-concurrency revision.",
            "minimum": 0
          },
          "software_id": {
            "type": "string",
            "description": "Parent software identity."
          },
          "state": {
            "type": "string",
            "description": "Highest achieved lifecycle state."
          },
          "version": {
            "type": "string",
            "description": "Exact opaque upstream version."
          }
        }
      },
...
      "SoftwareStatusResponse": {
        "type": "object",
        "description": "Bounded operator view of execution and publication for one software item.",
        "required": [
          "software_id",
          "channels",
          "enabled_targets",
          "outstanding_runs",
          "blocked_targets"
        ],
        "properties": {
          "blocked_targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockedTargetResponse"
            },
            "description": "First 200 enabled targets without a matching worker observed in the last five minutes."
          },
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SoftwareChannelStatusResponse"
            },
            "description": "Current published channels."
          },
          "enabled_targets": {
            "type": "integer",
            "format": "int64",
            "description": "Enabled target count.",
            "minimum": 0
          },
          "last_success_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Last successful check, including no-change observations."
          },
          "latest_run": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RunSummaryResponse",
                "description": "Latest run observation."
              }
            ]
          },
          "next_run_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Next recurring check."
          },
          "outstanding_runs": {
            "type": "integer",
            "format": "int64",
            "description": "Queued and running builds.",
            "minimum": 0
          },
          "software_id": {
            "type": "string",
            "description": "Software identity."
          }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "description": "One field-level validation failure.",
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable validation code."
          },
          "field": {
            "type": "string",
            "description": "Request field or parameter."
          },
          "message": {
            "type": "string",
            "description": "Human-readable diagnostic."
          }
        }
      },
...
      "WorkerResponse": {
        "type": "object",
        "description": "Administrative worker status and capabilities.",
        "required": [
          "draining",
          "id",
          "name",
          "allowed_capabilities",
          "advertised_capabilities",
          "enabled",
          "last_seen_at",
          "revision"
        ],
        "properties": {
          "advertised_capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Latest detected advertisement."
          },
          "allowed_capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Server-defined capability ceiling."
          },
          "draining": {
            "type": "boolean",
            "description": "Whether new claims are paused while existing attempts finish."
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the worker can authenticate."
          },
          "id": {
            "type": "string",
            "description": "UUIDv7 worker identity."
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "description": "Latest registration activity."
          },
          "name": {
            "type": "string",
            "description": "Operator-facing name."
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "description": "Optimistic-concurrency revision.",
            "minimum": 0
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}